CorsairLed main.cpp, patch 18, test

+++ CorsairLed.h
@@ -11,9 +11,9 @@ class CorsairLed {
 		~CorsairLed();
 		
 		struct CorsairLedColor {
-			uint8_t red;
-			uint8_t green;
-			uint8_t blue;
+			int red;
+			int green;
+			int blue;
 		} color[4];
 		
 		int GetLedCount();
